2020-01-07 Command 검증, D, ClassDAO, ReviewDAO 수정

+++ src/command/AdminReviewDeleteOkCommand.java
@@ -16,11 +16,13 @@ public class AdminReviewDeleteOkCommand implements Command {
 		AdminReviewDAO dao = new AdminReviewDAO();
 		int cnt = 0;
 		
-		try {
-			cnt = dao.deleteReview(review_uid);
-			request.setAttribute("adminReviewDeleteOk", cnt);
-		} catch (SQLException e) {
-			e.printStackTrace();
+		if (review_uid > 0) {
+			try {
+				cnt = dao.deleteReview(review_uid);
+				request.setAttribute("adminReviewDeleteOk", cnt);
+			} catch (SQLException e) {
+				e.printStackTrace();
+			}
 		}
 	}
